# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification - Transporte Escolar Eventos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locais:
    get:
      operationId: listLocais
      tags: [locais]
      summary: Listar locais de eventos
      responses:
        "200":
          description: Lista de locais
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Local"
    post:
      operationId: createLocal
      tags: [locais]
      summary: Cadastrar local de evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalInput"
      responses:
        "201":
          description: Local criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Local"
  /locais/{id}:
    get:
      operationId: getLocal
      tags: [locais]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Local
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Local"
    patch:
      operationId: updateLocal
      tags: [locais]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalUpdate"
      responses:
        "200":
          description: Local atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Local"
    delete:
      operationId: deleteLocal
      tags: [locais]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado

  # ── Tempos de Deslocamento ──────────────────────────────────────────────────
...
  /eventos:
    get:
      operationId: listEventos
      tags: [eventos]
      summary: Listar eventos
      parameters:
        - name: local_id
          in: query
          schema: { type: integer }
        - name: data_inicio
          in: query
          schema: { type: string, format: date }
        - name: data_fim
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evento"
    post:
      operationId: createEvento
      tags: [eventos]
      summary: Criar evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoInput"
      responses:
        "201":
          description: Evento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evento"
  /eventos/{id}:
    get:
      operationId: getEvento
      tags: [eventos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Evento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evento"
    patch:
      operationId: updateEvento
      tags: [eventos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventoUpdate"
      responses:
        "200":
          description: Evento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evento"
    delete:
      operationId: deleteEvento
      tags: [eventos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado
  /eventos/{id}/resumo:
    get:
      operationId: getEventoResumo
      tags: [eventos]
      summary: Resumo do evento - alunos, vagas, veículos
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Resumo do evento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventoResumo"

  # ── Sessões ─────────────────────────────────────────────────────────────────
  /eventos/{eventoId}/sessoes:
    get:
      operationId: listSessoes
      tags: [sessoes]
      summary: Listar sessões de um evento
      parameters:
        - name: eventoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lista de sessões
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sessao"
    post:
      operationId: createSessao
      tags: [sessoes]
      summary: Criar sessão para um evento
      parameters:
        - name: eventoId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessaoInput"
      responses:
        "201":
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sessao"
  /eventos/{eventoId}/sessoes/{id}:
    patch:
      operationId: updateSessao
      tags: [sessoes]
      parameters:
        - name: eventoId
          in: path
          required: true
          schema: { type: integer }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessaoUpdate"
      responses:
        "200":
          description: Sessão atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sessao"
    delete:
      operationId: deleteSessao
      tags: [sessoes]
      parameters:
        - name: eventoId
          in: path
          required: true
          schema: { type: integer }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado

  # ── Escalas ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Local:
      type: object
      required: [id, nome]
      properties:
        id:       { type: integer }
        nome:     { type: string }
        endereco: { type: ["string", "null"] }
        cidade:   { type: ["string", "null"] }
        observacoes: { type: ["string", "null"] }
    LocalInput:
      type: object
      required: [nome]
      properties:
        nome:        { type: string }
        endereco:    { type: string }
        cidade:      { type: string }
        observacoes: { type: string }
    LocalUpdate:
      type: object
      properties:
        nome:        { type: string }
        endereco:    { type: string }
        cidade:      { type: string }
        observacoes: { type: string }

    # ── Tempo de Deslocamento ─────────────────────────────────────────────────
...
    Evento:
      type: object
      required: [id, nome, data, local_id]
      properties:
        id:          { type: integer }
        nome:        { type: string }
        data:        { type: string, format: date }
        local_id:    { type: integer }
        local_nome:  { type: ["string", "null"] }
        observacoes: { type: ["string", "null"] }
        total_sessoes: { type: ["integer", "null"] }
    EventoInput:
      type: object
      required: [nome, data, local_id]
      properties:
        nome:        { type: string }
        data:        { type: string, format: date }
        local_id:    { type: integer }
        observacoes: { type: string }
    EventoUpdate:
      type: object
      properties:
        nome:        { type: string }
        data:        { type: string, format: date }
        local_id:    { type: integer }
        observacoes: { type: string }
    EventoResumo:
      type: object
      required: [evento_id, total_alunos, total_vagas, vagas_sobra, total_veiculos, sessoes]
      properties:
        evento_id:      { type: integer }
        evento_nome:    { type: string }
        total_alunos:   { type: integer }
        total_vagas:    { type: integer }
        vagas_sobra:    { type: integer }
        total_veiculos: { type: integer }
        sessoes:
          type: array
          items:
            $ref: "#/components/schemas/SessaoResumo"

    # ── Sessão ────────────────────────────────────────────────────────────────
    Sessao:
      type: object
      required: [id, evento_id, horario_inicio, duracao_minutos]
      properties:
        id:               { type: integer }
        evento_id:        { type: integer }
        nome:             { type: ["string", "null"] }
        horario_inicio:   { type: string }
        duracao_minutos:  { type: integer }
        horario_termino:  { type: ["string", "null"] }
        periodo_sessao:   { type: ["string", "null"] }
    SessaoInput:
      type: object
      required: [horario_inicio, duracao_minutos]
      properties:
        nome:            { type: string }
        horario_inicio:  { type: string }
        duracao_minutos: { type: integer }
    SessaoUpdate:
      type: object
      properties:
        nome:            { type: string }
        horario_inicio:  { type: string }
        duracao_minutos: { type: integer }
    SessaoResumo:
      type: object
      required: [sessao_id, horario_inicio, horario_termino, total_escalas]
      properties:
        sessao_id:       { type: integer }
        nome:            { type: ["string", "null"] }
        horario_inicio:  { type: string }
        horario_termino: { type: string }
        total_escalas:   { type: integer }
        total_alunos:    { type: integer }
        total_vagas:     { type: integer }

    # ── Escala ────────────────────────────────────────────────────────────────
...