# lib/api-zod/src/generated/types/eventoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Transporte Escolar Eventos
 * OpenAPI spec version: 0.1.0
 */

export interface EventoInput {
  nome: string;
  data: Date;
  local_id: number;
  observacoes?: string;
}