# lib/api-zod/src/generated/types/sessaoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Transporte Escolar Eventos
 * OpenAPI spec version: 0.1.0
 */

export interface SessaoUpdate {
  nome?: string;
  horario_inicio?: string;
  duracao_minutos?: number;
}